Consolidate .netcoreapp.cs files because System.Security.Cryptography.* projects is no longer cross-compiled
